Ultimos resultados antes do merge NPprocedure

+++ MINAS-BR_jan2021/src/br/OfflinePhase.java
@@ -278,7 +278,7 @@ public final class OfflinePhase{
             this.model.incrementNumerOfObservedExamples();
         }
         this.trainingData = trainingData;
-        this.model.setCurrentCardinality(Math.ceil(qtdeRotulos/D.size()));
+        this.model.setCurrentCardinality(Math.ceil((double)qtdeRotulos/(double)D.size()));
     }
